Changed score to float64 from int.  Updated README.md

+++ cmd/npuzzle/score.go
@@ -1,16 +1,16 @@
 package main
 
 // breadth first search
-func Uniform(state *State) int {
+func Uniform(state *State) float64 {
 	return state.Dist
 }
 
 // heuristic depth first search
-func Greedy(state *State) int {
+func Greedy(state *State) float64 {
 	return state.H
 }
 
 // CCC-Combo Breaker!!
-func Astar(state *State) int {
+func Astar(state *State) float64 {
 	return state.Dist + state.H
 }
